+++ db/flush.go
@@ -10,12 +10,13 @@ import (
 	"go.uber.org/zap"
 )
 
-func (l *Loader) Flush(ctx context.Context, outputModuleHash string, cursor *sink.Cursor) (err error) {
+func (l *Loader) Flush(ctx context.Context, outputModuleHash string, cursor *sink.Cursor) (rowFlushedCount int, err error) {
 	ctx = clickhouse.Context(context.Background(), clickhouse.WithStdAsync(false))
+
 	startAt := time.Now()
 	tx, err := l.DB.BeginTx(ctx, nil)
 	if err != nil {
-		return fmt.Errorf("failed to being db transaction: %w", err)
+		return 0, fmt.Errorf("failed to being db transaction: %w", err)
 	}
 	defer func() {
 		if err != nil {
@@ -24,22 +25,25 @@ func (l *Loader) Flush(ctx context.Context, outputModuleHash string, cursor *sin
 			}
 		}
 	}()
-	entryCount, err := l.getDialect().Flush(tx, ctx, l, outputModuleHash, cursor)
+
+	rowFlushedCount, err = l.getDialect().Flush(tx, ctx, l, outputModuleHash, cursor)
 	if err != nil {
-		return fmt.Errorf("dialect flush: %w", err)
+		return 0, fmt.Errorf("dialect flush: %w", err)
 	}
-	entryCount += 1
+
+	rowFlushedCount += 1
 	if err := l.UpdateCursor(ctx, tx, outputModuleHash, cursor); err != nil {
-		return fmt.Errorf("update cursor: %w", err)
+		return 0, fmt.Errorf("update cursor: %w", err)
 	}
 
 	if err := tx.Commit(); err != nil {
-		return fmt.Errorf("failed to commit db transaction: %w", err)
+		return 0, fmt.Errorf("failed to commit db transaction: %w", err)
 	}
 	l.reset()
 
-	l.logger.Debug("flushed table(s) entries", zap.Int("table_count", l.entries.Len()), zap.Int("entry_count", entryCount), zap.Duration("took", time.Since(startAt)))
-	return nil
+	// We add + 1 to the table count because the `cursors` table is an implicit table
+	l.logger.Debug("flushed table(s) rows to database", zap.Int("table_count", l.entries.Len()+1), zap.Int("row_count", rowFlushedCount), zap.Duration("took", time.Since(startAt)))
+	return rowFlushedCount, nil
 }
 
 func (l *Loader) reset() {